DP-6093-BIRD-writes-nexthop-255-255-255-255
