+++ proto/mrt/mrt.c
@@ -730,10 +730,15 @@ mrt_table_dump_free(struct mrt_table_dump_state *s)
 static int
 mrt_table_dump_step(struct mrt_table_dump_state *s)
 {
-  struct bgp_write_state bws = { .as4_session = 1 };
-
-  s->max = 2048;
-  s->bws = &bws;
+  if (s->table_open) {
+    s->max = 2048;
+    goto step;
+  }
+  else {
+    struct bgp_write_state bws = { .as4_session = 1 };
+    s->max = 2048;
+    s->bws = &bws;
+  }
 
   if (s->table_open)
     goto step;
